+++ src/main/java/abyssredemption/network/ProtocolConstants.java
@@ -1,11 +1,14 @@
 package abyssredemption.network;
 
 public final class ProtocolConstants {
-    public static final int PROTOCOL_VERSION = 1;
-    public static final int CAP_PLAYTIME = 1;
-    public static final int CAP_DEATHS = 1 << 1;
-    public static final int CAP_BLOCKS_TOTAL = 1 << 2;
-    public static final int CAP_BLOCKS_WEEKLY = 1 << 3;
-    public static final int ALL_CAPABILITIES = CAP_PLAYTIME | CAP_DEATHS | CAP_BLOCKS_TOTAL | CAP_BLOCKS_WEEKLY;
+    public static final int PROTOCOL_VERSION = 2;
+    public static final int CAP_OVERVIEW = 1;
+    public static final int CAP_ACTIVITY = 1 << 1;
+    public static final int CAP_PLAYTIME = 1 << 2;
+    public static final int CAP_DEATHS = 1 << 3;
+    public static final int CAP_PLACEMENTS = 1 << 4;
+    public static final int CAP_DAILY_TREND = 1 << 5;
+    public static final int CAP_WEEKLY_TREND = 1 << 6;
+    public static final int ALL_CAPABILITIES = CAP_OVERVIEW | CAP_ACTIVITY | CAP_PLAYTIME | CAP_DEATHS | CAP_PLACEMENTS | CAP_DAILY_TREND | CAP_WEEKLY_TREND;
     private ProtocolConstants() {}
 }